Move mdx-page to basic
