+++ packages/integrations/mdx/test/fixtures/basic/src/pages/page/_styles.css
