+++ packages/integrations/mdx/test/fixtures/basic/src/pages/page/index.mdx
@@ -1,0 +1,3 @@
+import './_styles.css'
+
+# Hello page!
