--- packages/integrations/mdx/test/fixtures/mdx-page/astro.config.ts
@@ -1,5 +1,0 @@
-import mdx from '@astrojs/mdx';
-
-export default {
-	integrations: [mdx()]
-}
