--- packages/integrations/mdx/test/fixtures/mdx-page/src/pages/index.mdx
@@ -1,3 +1,0 @@
-import '../styles.css'
-
-# Hello page!
